backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic==1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0,>=3.1.0
python-multipart==0.0.6
httpx==0.25.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
reportlab==4.0.7
openai==1.3.5
python-dotenv==1.0.0
python-dateutil==2.8.2
# Email
aiosmtplib==3.0.1
email-validator==2.1.0
# DOCX export
python-docx==1.1.0
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0
